Add files via upload
Logistics in supply chain plays a crucial role in the inventory management of the raw materials required for the operations, and a slight delay in the delivery of these goods would cost the company heavily in terms of brand value and customer satisfaction if they fail to meet their committed delivery schedule. Hence it becomes critical for any supply chain organization to proactively manage the logistics to deliver the goods on time.
The Logistics firm wants to use a data driven strategy to manage its operations. It has fetched some historic data of the orders placed. They want to train an ML model to predict the delivery status of an order based on various parameters. They want to use this model to predict which orders might get delayed in the future and hence take preventive measures.